Add the very very very lose code for this project
Hi Mr. Oswald!
I woke up this morning and wanted to check on a few things before I submitted but I saw all of my code deleted or wasn't uploaded correctly to GitHub as I was in a place coding this with not great WiFi over the week so I am now trying to upload everything again. Luckily, through my coding process, I made different files for each of my iterations on P5 because I wanted to be able to go back and see so I am able to do that now! Will also send you an email about this.